% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbvnorm}
\alias{dbvnorm}
\title{Bivariate normal density}
\usage{
dbvnorm(y, mu, sigma, log_lik = FALSE)
}
\arguments{
\item{y}{vector of length 2 of observations}

\item{mu}{vector of length 2 of means}

\item{sigma}{2 x 2 variance-covariance matrix}

\item{log_lik}{bool}
}
\value{
evaluated density
}
\description{
density of bivariate normal distribution
}
