% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvnorm_arma}
\alias{dmvnorm_arma}
\title{Multivariate normal density}
\usage{
dmvnorm_arma(x, mean, sigma, log = FALSE)
}
\arguments{
\item{x}{matrix of data}

\item{mean}{vector of means for MVN density}

\item{sigma}{variance-covariance matrix for MVN density}
}
\value{
logretval evaluated density
}
\description{
from http://gallery.rcpp.org/articles/dmvnorm_arma/
}
