% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_causal.R
\name{estimate_lambda_rho_given_pi}
\alias{estimate_lambda_rho_given_pi}
\title{Estimate lambda and rho}
\usage{
estimate_lambda_rho_given_pi(Y, S, U, pi_vec)
}
\arguments{
\item{Y}{matrix of betahats n x 2}

\item{S}{matrix of std. errors n x 2}

\item{U}{array prior diagonal 2 x 2 covariance matricies}

\item{pi_vec}{vector prior K mixture proportions (weights)}
}
\value{
lambda_rho list of estimated parameters
}
\description{
estimate lambda and rho condtional on 
knowing the mixture proportions using 
a blackbox numerical optimizer
}
