% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_bvash.R
\name{estimate_lambda_rho_given_w}
\alias{estimate_lambda_rho_given_w}
\title{Estimate lambda and rho}
\usage{
estimate_lambda_rho_given_w(Y, w, U, S)
}
\arguments{
\item{Y}{matrix of betahats n x 2}

\item{w}{vector prior K mixture proportions (weights)}

\item{U}{array prior diagonal 2 x 2 covariance matricies}

\item{S}{matrix of std. errors n x 2}
}
\value{
lambda_rho list of estimated parameters
}
\description{
estimate lambda and rho condtional on 
             knowing the mixture proportions using 
             a numerical optimizer
}
