% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_causal.R
\name{estimate_pi_given_lambda_rho}
\alias{estimate_pi_given_lambda_rho}
\title{Estimate the mixture proportions}
\usage{
estimate_pi_given_lambda_rho(Y, S, U, lambda_12, lambda_21, rho)
}
\arguments{
\item{Y}{matrix of betahats n x 2}

\item{S}{matrix of std. errors n x 2}

\item{U}{array prior diagonal 2 x 2 covariance matricies}

\item{lambda_12}{double causal effect of trait 2 --> trait 1}

\item{lambda_21}{double causal effect of trait 1 --> trait 2}

\item{rho}{double correlation of the estimated effects}
}
\value{
pi_vec vector next estimate of mixture proportions
}
\description{
estimates the mixture proportions via convex optimzation condtional
on the current estimate of the loadings matrix
}
