% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_causal.R
\name{estimate_pi_lambda_rho}
\alias{estimate_pi_lambda_rho}
\title{Estimate pi, lambda, and rho in 2-step iterative algorithim}
\usage{
estimate_pi_lambda_rho(Y, S, U, max_iter = 10)
}
\arguments{
\item{Y}{matrix of betahats n x 2}

\item{S}{matrix of std. errors n x 2}

\item{U}{array prior diagonal 2 x 2 covariance matricies}
}
\value{
pi_lambda_rho_res list of optimized parameters
}
\description{
estimates pi, lambda and rho in a 2-step iterative algorithim where
in each iteration we first estimate w given lambda and rho by convex
optimazation and then estimate lambda and rho given pi using black box 
numerical optimization
}
