% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvashr.R
\name{estimate_prior}
\alias{estimate_prior}
\title{Estimate prior}
\usage{
estimate_prior(X, S, K)
}
\arguments{
\item{X}{n x 2 matrix of estimated effects}

\item{S}{n x 2 matrix of std. errors}

\item{K}{number of mixture components to use in EM}
}
\value{
prior_res list from output of ExtremeDeconvolution
}
\description{
TODO: fill in
}
