% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_sharing.R
\name{estimate_sign_pi}
\alias{estimate_sign_pi}
\title{Compute sign mixture proportions}
\usage{
estimate_sign_pi(Y, S)
}
\arguments{
\item{Y}{matrix of betahats n x 2}

\item{S}{matrix of std. errors n x 2}
}
\value{
pi_df data.frame of sign mixture probabilities
}
\description{
TODO: description
}
