% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{get_causal_comp_lik_i}
\alias{get_causal_comp_lik_i}
\title{Get an observation's likelihood under causal model}
\usage{
get_causal_comp_lik_i(y_i, lambda_12, lambda_21, U, rho, s_1, s_2)
}
\arguments{
\item{y_i}{vector containing betahats for two traits}

\item{lambda_12}{double causal effect of trait 2 --> trait 1}

\item{lambda_21}{double causal effect of trait 1 --> trait 2}

\item{U}{array prior diagonal 2 x 2 covariance matricies}

\item{rho}{double correlation of estimated effects}

\item{s_1}{double std. error for trait 1}

\item{s_2}{double std. error for trait 2}
}
\value{
lik_i vector likelihood under each component for a single observation
}
\description{
computes the likelihood for a single observation
             for the bvash causal model
}
