% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{get_causal_lik_mat}
\alias{get_causal_lik_mat}
\title{Get likelihood matrix under causal model}
\usage{
get_causal_lik_mat(Y, lambda_12, lambda_21, U, rho, S)
}
\arguments{
\item{Y}{matrix of betahats n x 2}

\item{lambda_12}{double causal effect of trait 2 --> trait 1}

\item{lambda_21}{double causal effect of trait 1 --> trait 2}

\item{U}{array prior diagonal 2 x 2 covariance matricies}

\item{rho}{double correlation of the estimated effects}

\item{S}{matrix of std. errors n x 2}
}
\value{
L matrix likelihood matrix n x K
}
\description{
computes likelihood matrix under the bvash causal model
}
