% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_lik_mat_cfm}
\alias{get_lik_mat_cfm}
\title{Gets likelihood matrix for causal factor model (cfm)}
\usage{
get_lik_mat_cfm(y, s, u, lambda_12, lambda_21, rho)
}
\arguments{
\item{y}{matrix n x 2 of observations}

\item{s}{matrix n x 2 of std. errors}

\item{u}{cube 2 x 2 x K of prior covariance matricies}

\item{lambda_12}{double effect of varialbe 2 on variable 1}

\item{lambda_21}{double effect of variable 1 on variable 2}

\item{rho}{double correlation of estimated effects}
}
\value{
lik_mat matrix n x K of component likelihoods
}
\description{
Computes n x K matrix of component likelihoods
}
