% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_lik_mat_ss}
\alias{get_lik_mat_ss}
\title{Compute likelihood matrix for signing sharing model}
\usage{
get_lik_mat_ss(m)
}
\arguments{
\item{m}{matrix of sign probabilites}
}
\description{
TODO: description
}
