% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_log_likelihood_cfm}
\alias{get_log_likelihood_cfm}
\title{Gets log likelihood for causal factor model (cfm)}
\usage{
get_log_likelihood_cfm(y, s, u, lambda_12, lambda_21, rho, pi_vec)
}
\arguments{
\item{y}{matrix n x 2 of observations}

\item{s}{matrix n x 2 of std. errors}

\item{u}{cube 2 x 2 x K of prior covariance matricies}

\item{lambda_12}{double effect of varialbe 2 on variable 1}

\item{lambda_21}{double effect of variable 1 on variable 2}

\item{rho}{double correlation of estimated effects}

\item{pi_vec}{vector vector of mixture proportions}
}
\value{
ll double likelihood under causal factor model
}
\description{
Computes likelihood under causal factor model
}
