% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bv.R
\name{get_prior_comp}
\alias{get_prior_comp}
\title{Get prior component parameters}
\usage{
get_prior_comp(prior_res, table = FALSE)
}
\arguments{
\item{prior_res}{list output from estimate_prior}

\item{table}{boolean if true return kable defaults to False}
}
\value{
parameters of prior components
}
\description{
computes formatted table (df, or kable) of fitted prior
             paramters including mixture propotions, correlation coef, 
             and std devs
}
