% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvashr.R
\name{get_prior_comp_df}
\alias{get_prior_comp_df}
\title{Get prior component parameters}
\usage{
get_prior_comp_df(prior_res, table = FALSE)
}
\arguments{
\item{prior_res}{list output from extreme deconvolution}

\item{table}{boolean if true return kable defaults to False}
}
\value{
parameters of prior components
}
\description{
TODO: fill in
}
