% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_sharing.R
\name{get_sign_prob_mat}
\alias{get_sign_prob_mat}
\title{Compute matrix of sign probabilities}
\usage{
get_sign_prob_mat(Y, S)
}
\arguments{
\item{Y}{matrix of betahats n x 2}

\item{S}{matrix of std. errors n x 2}
}
\value{
M matrix n x 6 of sign probabilities
}
\description{
computes matrix of of sign probabilities for 
a pair of traits
}
