% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log sum exponential}
\usage{
log_sum_exp(y)
}
\arguments{
\item{y}{vector to be log-sum-exponentiated}
}
\value{
lse doube of log-sum-exponentiated vector
}
\description{
inspired by the implementation in https://github.com/dcgerard/updog/blob/master/src/utitility.cpp
}
