% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mahalanobis}
\alias{Mahalanobis}
\title{Mahalanobis distance}
\usage{
Mahalanobis(x, center, cov)
}
\arguments{
\item{x}{matrix of data}

\item{mean}{vector of means for MVN density}

\item{sigma}{variance-covariance matrix for MVN density}
}
\value{
mahalanobis distance
}
\description{
from http://gallery.rcpp.org/articles/dmvnorm_arma/
}
