% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bv.R
\name{plot_prior}
\alias{plot_prior}
\title{Plot prior}
\usage{
plot_prior(trait_1, trait_2, prior_res)
}
\arguments{
\item{trait_1}{char first trait name}

\item{trait_2}{char second trait name}

\item{prior_res}{list output from estimate_prior}
}
\value{
p ggplot object of prior plot
}
\description{
plots ellipes of fitted prior covariance
             matricies of each mixture component
}
