% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvashr.R
\name{sample_prior}
\alias{sample_prior}
\title{Sample prior}
\usage{
sample_prior(N, prior_res)
}
\arguments{
\item{N}{number of samples}

\item{prior_res}{list output from extreme deconvolution}
}
\value{
R N x 3 matrix of samples and latent component as columns
}
\description{
TODO: fill in
}
