% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bv.R
\name{se_mat_to_arr}
\alias{se_mat_to_arr}
\title{Convert std. error matrix to array}
\usage{
se_mat_to_arr(S, rho)
}
\value{
S_arr array of estimated effect error covariance matricies
}
\description{
converts std. error matrix (n x 2) into 
             array (2 x 2 x n) that in includes that null
             correlation on the off diagonal
}
